+++ spec/config_parser_spec.rb
@@ -12,10 +12,10 @@ describe Makara::ConfigParser do
             :name => 'themaster'
           },
           {
-            :name => 'slave1'
+            :name => 'replica1'
           },
           {
-            :name => 'slave2'
+            :name => 'replica2'
           }
         ]
       }
@@ -108,8 +108,8 @@ describe Makara::ConfigParser do
     expect(parser.id).to eq('myproxyidwithreservedcharacters')
   end
 
-  context 'master and slave configs' do
-    it 'should provide master and slave configs' do
+  context 'master and replica configs' do
+    it 'should provide master and replica configs' do
       parser = described_class.new(config)
       expect(parser.master_configs).to eq([
         {
@@ -120,16 +120,16 @@ describe Makara::ConfigParser do
           :master_ttl => 5
         }
       ])
-      expect(parser.slave_configs).to eq([
+      expect(parser.replica_configs).to eq([
         {
-          :name => 'slave1',
+          :name => 'replica1',
           :top_level => 'value',
           :sticky => true,
           :blacklist_duration => 30,
           :master_ttl => 5
         },
         {
-          :name => 'slave2',
+          :name => 'replica2',
           :top_level => 'value',
           :sticky => true,
           :blacklist_duration => 30,
@@ -141,7 +141,7 @@ describe Makara::ConfigParser do
     it 'connection configuration should override makara config' do
       config[:makara][:blacklist_duration] = 123
       config[:makara][:connections][0][:blacklist_duration] = 456
-      config[:makara][:connections][1][:top_level] = 'slave value'
+      config[:makara][:connections][1][:top_level] = 'replica value'
 
       parser = described_class.new(config)
       expect(parser.master_configs).to eq([
@@ -153,16 +153,16 @@ describe Makara::ConfigParser do
           :master_ttl => 5
         }
       ])
-      expect(parser.slave_configs).to eq([
+      expect(parser.replica_configs).to eq([
         {
-          :name => 'slave1',
-          :top_level => 'slave value',
+          :name => 'replica1',
+          :top_level => 'replica value',
           :sticky => true,
           :blacklist_duration => 123,
           :master_ttl => 5
         },
         {
-          :name => 'slave2',
+          :name => 'replica2',
           :top_level => 'value',
           :sticky => true,
           :blacklist_duration => 123,